model diagram as figureM1
